Abstracted the response list animation in ChooseStock to it's own custom hook.
